fix popout bug in demo site
